CODE/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
pdfminer.six
pdfplumber
transformers
torch
